Broke beams_functors.hpp into many header files.
