+++ src/restruct_poc/beams_input.cpp
@@ -2,7 +2,12 @@
 
 #include <algorithm>
 
-#include "beams_functors.hpp"
+#include "InterpolateQPPosition.hpp"
+#include "InterpolateQPRotation.hpp"
+#include "CalculateJacobian.hpp"
+#include "InterpolateQPState.hpp"
+#include "InterpolateQPVelocity.hpp"
+#include "InterpolateQPAcceleration.hpp"
 
 namespace openturbine {
 
@@ -313,16 +318,24 @@ Beams CreateBeams(const BeamsInput& beams_input) {
             beams.node_u, beams.qp_u, beams.qp_u_prime, beams.qp_r, beams.qp_r_prime}
     );
     Kokkos::parallel_for(
-        "InterpolateQPVelocity", beams.num_elems,
-        InterpolateQPVelocity{
-            beams.elem_indices, beams.shape_interp, beams.shape_deriv, beams.qp_jacobian,
-            beams.node_u_dot, beams.qp_u_dot, beams.qp_omega}
+        "InterpolateQPVelocity", Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_qps}},
+        InterpolateQPVelocity_Translation{
+            beams.elem_indices, beams.shape_interp, beams.node_u_dot, beams.qp_u_dot}
     );
     Kokkos::parallel_for(
-        "InterpolateQPAcceleration", beams.num_elems,
-        InterpolateQPAcceleration{
-            beams.elem_indices, beams.shape_interp, beams.shape_deriv, beams.qp_jacobian,
-            beams.node_u_ddot, beams.qp_u_ddot, beams.qp_omega_dot}
+        "InterpolateQPVelocity", Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_qps}},
+        InterpolateQPVelocity_Angular{
+            beams.elem_indices, beams.shape_interp, beams.node_u_dot, beams.qp_omega}
+    );
+    Kokkos::parallel_for(
+        "InterpolateQPAcceleration", Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_qps}},
+        InterpolateQPAcceleration_Translation{
+            beams.elem_indices, beams.shape_interp, beams.node_u_ddot, beams.qp_u_ddot}
+    );
+    Kokkos::parallel_for(
+        "InterpolateQPAcceleration", Kokkos::MDRangePolicy{{0, 0}, {beams.num_elems, beams.max_elem_qps}},
+        InterpolateQPAcceleration_Angular{
+            beams.elem_indices, beams.shape_interp, beams.node_u_ddot, beams.qp_omega_dot}
     );
     return beams;
 }
